backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
redis==6.0.0
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
bcrypt==4.0.1
